refactor(shared): promote the player components out of the players domain
P2.5, second half and the bigger one. MAY_REACH_INSIDE 17 -> 9, and two
more cycles dissolved, 8 -> 6.

Eight domains reached into players/components/player for PlayerSummary and
PlayerSummaryPoints. Ten importers updated in total -- those eight plus
players/player.page.tsx and players/components/player-stats-table.tsx,
which owned it.

P2.1b was genuinely the unblock. Before it, player.tsx imported
EnhancedPlayerData from scoring/types, so moving the file would have added
new _shared -> domain edges and undone what P2.1b just finished. With
those types in the kernel the moved file imports only _shared paths --
zero domain imports -- and SHARED_MAY_IMPORT stays empty.

DisplayablePlayer travelled with the component rather than going into the
kernel. player.tsx was its only consumer and it is a component prop type,
so it now lives in the component file, per "component props live next to
their components". Putting it in the kernel would also have forced
player-types.ts to import squad-types.ts, breaking the sibling
relationship P2.1b just established between them.

Two cycles dissolved, not the one forecast. players <-> wishlist was
predicted: wishlist-details.tsx -> players/components/player was
wishlist's only outbound edge to players. The second falls out of
player-stats-table.tsx being a tenth importer nobody had counted.

That tenth importer was nearly missed: the grep used to find call sites
filtered out filenames containing player-stats-table, which hid its
`from './player'` sibling import. yarn typecheck caught it as +1 against
the 176 baseline -- the ratchet earning its keep again.

Verification: type errors 176, unchanged. 346 tests pass, ratchet counts
all unchanged, yarn build clean.

Also records the six remaining cycles against P2.6, two of which
(admin <-> draft, draft <-> scoring) were never on its original list.
