+++ draft/app/admin/components/sections/transfers-section.tsx
@@ -4,12 +4,12 @@ import type React from 'react';
 import { useEffect, useState } from 'react';
 import { useFetcher, useNavigate } from 'react-router';
 import { GameweekSelector } from '../../../_shared/components/gameweek-selector';
+import { PlayerSummary } from '../../../_shared/components/player';
 import { SelectDivision } from '../../../_shared/components/select-division';
 import { Table, type TableColumn } from '../../../_shared/components/table';
 import type { FplTeam, GameWeekData } from '../../../_shared/lib/fpl/fpl-types';
 import type { DivisionSheetData, ManagerId, PositionSlotKey } from '../../../_shared/types/league-types';
 import type { RosterPlayer } from '../../../_shared/types/squad-types';
-import { PlayerSummary } from '../../../players/components/player';
 import { LoanStatusDisplay } from '../../../transfers/components/loan-status-display';
 import type { TransferAdminOverviewData, TransferValidationResult } from '../../../transfers/types/transfer-rule-types';
 import type { ProcessedTransfer } from '../../../transfers/types/transfer-types';
