+++ draft/app/players/types/player-types.ts
@@ -3,24 +3,7 @@
 import type { FplPlayerSeasonFixture, FplTeam, GameWeekData } from '../../_shared/lib/fpl/fpl-types';
 import type { CustomPosition } from '../../_shared/types/league-types';
 import type { EnhancedPlayerData } from '../../_shared/types/player-types';
-import type { RosterPlayer } from '../../_shared/types/squad-types';
 import type { GameweekStatWithPoints, SeasonTotals } from '../../scoring/types/scoring-types';
-
-/**
- * A player as the shared player components render it.
- *
- * The same components display players from both sides of the app: an FPL player
- * (EnhancedPlayerData, from the player list, transfers and the wishlist) and a drafted
- * squad member (RosterPlayer, from a team sheet). They read every field with a fallback
- * -- `playerCode || code`, `playerName || web_name`, `playerPosition || draft.position`
- * -- so one of the two shapes is required and the other's fields are optional.
- *
- * This was previously typed as `EnhancedPlayerData & RosterPlayer`, requiring BOTH,
- * which no caller could ever satisfy -- hence the casts and the errors at every use.
- */
-export type DisplayablePlayer =
-    | (EnhancedPlayerData & Partial<RosterPlayer>)
-    | (RosterPlayer & Partial<EnhancedPlayerData>);
 
 // Map position codes to their full names
 export type PositionNameMap = {
