+++ draft/app/transfers/components/transfer-selector-stat-columns.tsx
@@ -1,9 +1,9 @@
 // app/transfers/components/transfer-selector-stat-columns.tsx
 
+import { PlayerSummary } from '../../_shared/components/player';
 import type { TableColumn } from '../../_shared/components/table';
 import type { FplTeam } from '../../_shared/lib/fpl/fpl-types';
 import type { EnhancedPlayerData } from '../../_shared/types/player-types';
-import { PlayerSummary } from '../../players/components/player';
 import { PointsBreakdownTooltip } from '../../scoring';
 import { isStatRelevant } from '../../scoring/lib';
